+++ changelog.md
@@ -3,6 +3,8 @@
 
 **pycite 0.1.1**
 
+* Added `split_authors` a simple method to clean and abbreviate author names. 
+
 * Fixed issues with actions not running on GitHub.
 
 * Updated documentation